[Graphics] Resolved runtime semaphore synchronization between framegraph passes;

+++ Engine/src/Razix/Platform/API/OpenGL/OpenGLFrameBuffer.cpp
@@ -22,7 +22,7 @@ namespace Razix {
             glBindFramebuffer(GL_FRAMEBUFFER, m_FBO);
 
             // Get and bind the attachments (multiple color or depth)
-            for (size_t i = 0; i < frameBufInfo.attachmentCount; i++) {
+            for (uint32_t i = 0; i < frameBufInfo.attachmentCount; i++) {
 
                 // Only bind color attachments
                 if (frameBufInfo.renderPass->getAttachmentTypes()[i].type != RZTexture::Type::COLOR_2D)
